backend/requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
google-cloud-aiplatform
python-dotenv
python-multipart
requests

